Corrección del manejo de sesión de usuario
