+++ app/templates/public/base.html
@@ -38,6 +38,15 @@
             </div>
             <div class="col-lg-6 text-center text-lg-right">
                 <div class="d-inline-flex align-items-center">
+                    {% if session.get('nombre') %}
+                    <div class="btn-group align-items-center">
+                        <h6 class="m-0 mr-2">Hola, {{ session.get('nombre') }}</h6>
+                        <button type="button" class="btn btn-sm btn-light dropdown-toggle" data-toggle="dropdown">Mi Cuenta</button>
+                        <div class="dropdown-menu dropdown-menu-right">
+                            <a class="dropdown-item" href="{{ url_for('logout') }}"type="button">Cerrar Sesión</a>
+                        </div>
+                    </div>
+                    {% else %}
                     <div class="btn-group">
                         <button type="button" class="btn btn-sm btn-light dropdown-toggle" data-toggle="dropdown">Mi Cuenta</button>
                         <div class="dropdown-menu dropdown-menu-right">
@@ -45,6 +54,7 @@
                             <a class="dropdown-item" href="{{ url_for('registro') }}" type="button">Registrarse</a>
                         </div>
                     </div>
+                    {% endif %}
                 </div>
                 <div class="d-inline-flex align-items-center d-block d-lg-none">
                     <a href="" class="btn px-0 ml-2">
@@ -110,15 +120,7 @@
                             </div>
                         </div>
                         <div class="navbar-nav ml-auto py-0 d-none d-lg-block">
-<<<<<<< HEAD:app/templates/public/base.html
-                            <a href="" class="btn px-0">
-                                <i class="fas fa-heart text-primary"></i>
-                                <span class="badge text-secondary border border-secondary rounded-circle" style="padding-bottom: 2px;">0</span>
-                            </a>
-                            <a href="{{ url_for('carrito')}}" class="btn px-0 ml-3">
-=======
                             <a href="{{ url_for('carrito') }}" class="btn px-0 ml-3">
->>>>>>> dev:app/templates/base.html
                                 <i class="fas fa-shopping-cart text-primary"></i>
                                 <span class="badge text-secondary border border-secondary rounded-circle" style="padding-bottom: 2px;">0</span>
                             </a>
@@ -144,6 +146,8 @@
     {% block productos %}
     {% endblock %}
     {% block misproductos %}
+    {% endblock %}
+    {% block pago %}
     {% endblock %}
     <!-- Footer Start -->
     <div class="container-fluid bg-dark text-secondary mt-5 pt-5">
