+++ app/templates/public/checkout.html
@@ -88,6 +88,8 @@
                         <p>{{seleccionado['nombre']}}</p>
                         <p>{{seleccionado['precio']}}</p>
                     </div>
+                    {% endfor %}
+                </div>
                 <div class="pt-2">
                     <div class="d-flex justify-content-between mt-2">
                         <h5>Total</h5>
@@ -103,7 +105,7 @@
                             <label for="paypal">Debito/Credito</label>
                         </div>
                     </div>
-                    <a class="btn btn-block btn-primary font-weight-bold py-3" href=""> Pagar</a>
+                    <a class="btn btn-block btn-primary font-weight-bold py-3" href='{{ url_for("pagar") }}'> Pagar</a>
                 </div>
             </div>
         </div>
