+++ app/templates/public/pago.html
@@ -1,4 +1,4 @@
-{% extends "base.html" %}
+{% extends "public/base.html" %}
 
 {% block pago %}
 <div class="container mt-5">
@@ -6,7 +6,7 @@
     <form id="formularioPago">
         <div class="mb-3">
             <label for="numero" class="form-label">Número de Tarjeta</label>
-            <input type="text" class="form-control" id="numero" placeholder="XXXX XXXX XXXX XXXX" required>
+            <input type="text" class="form-control" id="numero" placeholder="XXXX XXXX XXXX XXXX" required maxlength="16">
         </div>
         <div class="mb-3">
             <label for="nombre" class="form-label">Nombre en la Tarjeta</label>
